@@ -1,9 +1,16 @@
 part of 'settings_page.dart';
 
 class ReaderSettings extends StatefulWidget {
-  const ReaderSettings({super.key, this.onChanged});
+  const ReaderSettings({
+    super.key,
+    this.onChanged,
+    this.comicId,
+    this.comicSource,
+  });
 
   final void Function(String key)? onChanged;
+  final String? comicId;
+  final String? comicSource;
 
   @override
   State<ReaderSettings> createState() => _ReaderSettingsState();
@@ -12,15 +19,57 @@ class ReaderSettings extends StatefulWidget {
 class _ReaderSettingsState extends State<ReaderSettings> {
   @override
   Widget build(BuildContext context) {
+    final comicId = widget.comicId;
+    final sourceKey = widget.comicSource;
+    final key = "$comicId@$sourceKey";
+
+    bool isEnabledSpecificSettings =
+        comicId != null &&
+        appdata.settings.isComicSpecificSettingsEnabled(comicId, sourceKey);
+
     return SmoothCustomScrollView(
       slivers: [
         SliverAppbar(title: Text("Reading".tl)),
+        if (comicId != null && sourceKey != null)
+          SliverMainAxisGroup(
+            slivers: [
+              SwitchListTile(
+                title: Text("Enable comic specific settings".tl),
+                value: isEnabledSpecificSettings,
+                onChanged: (b) {
+                  setState(() {
+                    appdata.settings.setEnabledComicSpecificSettings(
+                      comicId,
+                      sourceKey,
+                      b,
+                    );
+                  });
+                },
+              ).toSliver(),
+              if (isEnabledSpecificSettings)
+                Center(
+                  child: TextButton(
+                    onPressed: () {
+                      setState(() {
+                        appdata.settings.resetComicReaderSettings(key);
+                      });
+                    },
+                    child: Text(
+                      "Clear specific reader settings for this comic".tl,
+                    ),
+                  ),
+                ).toSliver(),
+              Divider().toSliver(),
+            ],
+          ),
         _SwitchSetting(
           title: "Tap to turn Pages".tl,
           settingKey: "enableTapToTurnPages",
           onChanged: () {
             widget.onChanged?.call("enableTapToTurnPages");
           },
+          comicId: isEnabledSpecificSettings ? widget.comicId : null,
+          comicSource: isEnabledSpecificSettings ? widget.comicSource : null,
         ).toSliver(),
         _SwitchSetting(
           title: "Reverse tap to turn Pages".tl,
@@ -28,6 +77,8 @@ class _ReaderSettingsState extends State<ReaderSettings> {
           onChanged: () {
             widget.onChanged?.call("reverseTapToTurnPages");
           },
+          comicId: isEnabledSpecificSettings ? widget.comicId : null,
+          comicSource: isEnabledSpecificSettings ? widget.comicSource : null,
         ).toSliver(),
         _SwitchSetting(
           title: "Page animation".tl,
@@ -35,6 +86,8 @@ class _ReaderSettingsState extends State<ReaderSettings> {
           onChanged: () {
             widget.onChanged?.call("enablePageAnimation");
           },
+          comicId: isEnabledSpecificSettings ? widget.comicId : null,
+          comicSource: isEnabledSpecificSettings ? widget.comicSource : null,
         ).toSliver(),
         SelectSetting(
           title: "Reading mode".tl,
@@ -58,6 +111,8 @@ class _ReaderSettingsState extends State<ReaderSettings> {
             }
             widget.onChanged?.call("readerMode");
           },
+          comicId: isEnabledSpecificSettings ? widget.comicId : null,
+          comicSource: isEnabledSpecificSettings ? widget.comicSource : null,
         ).toSliver(),
         _SliderSetting(
           title: "Auto page turning interval".tl,
@@ -69,6 +124,8 @@ class _ReaderSettingsState extends State<ReaderSettings> {
             setState(() {});
             widget.onChanged?.call("autoPageTurningInterval");
           },
+          comicId: isEnabledSpecificSettings ? widget.comicId : null,
+          comicSource: isEnabledSpecificSettings ? widget.comicSource : null,
         ).toSliver(),
         SliverAnimatedVisibility(
           visible: appdata.settings['readerMode']!.startsWith('gallery'),
@@ -84,6 +141,8 @@ class _ReaderSettingsState extends State<ReaderSettings> {
               setState(() {});
               widget.onChanged?.call("readerScreenPicNumberForLandscape");
             },
+            comicId: isEnabledSpecificSettings ? widget.comicId : null,
+            comicSource: isEnabledSpecificSettings ? widget.comicSource : null,
           ),
         ),
         SliverAnimatedVisibility(
@@ -99,10 +158,13 @@ class _ReaderSettingsState extends State<ReaderSettings> {
             onChanged: () {
               widget.onChanged?.call("readerScreenPicNumberForPortrait");
             },
+            comicId: isEnabledSpecificSettings ? widget.comicId : null,
+            comicSource: isEnabledSpecificSettings ? widget.comicSource : null,
           ),
         ),
         SliverAnimatedVisibility(
-          visible: appdata.settings['readerMode']!.startsWith('gallery') &&
+          visible:
+              appdata.settings['readerMode']!.startsWith('gallery') &&
               (appdata.settings['readerScreenPicNumberForLandscape'] > 1 ||
                   appdata.settings['readerScreenPicNumberForPortrait'] > 1),
           child: _SwitchSetting(
@@ -111,6 +173,23 @@ class _ReaderSettingsState extends State<ReaderSettings> {
             onChanged: () {
               widget.onChanged?.call("showSingleImageOnFirstPage");
             },
+            comicId: isEnabledSpecificSettings ? widget.comicId : null,
+            comicSource: isEnabledSpecificSettings ? widget.comicSource : null,
+          ),
+        ),
+        SliverAnimatedVisibility(
+          visible: appdata.settings['readerMode']!.startsWith('continuous'),
+          child: _SliderSetting(
+            title: "Mouse scroll speed".tl,
+            settingsIndex: "readerScrollSpeed",
+            interval: 0.1,
+            min: 0.5,
+            max: 3,
+            onChanged: () {
+              widget.onChanged?.call("readerScrollSpeed");
+            },
+            comicId: isEnabledSpecificSettings ? widget.comicId : null,
+            comicSource: isEnabledSpecificSettings ? widget.comicSource : null,
           ),
         ),
         _SwitchSetting(
@@ -120,6 +199,8 @@ class _ReaderSettingsState extends State<ReaderSettings> {
             setState(() {});
             widget.onChanged?.call('enableDoubleTapToZoom');
           },
+          comicId: isEnabledSpecificSettings ? widget.comicId : null,
+          comicSource: isEnabledSpecificSettings ? widget.comicSource : null,
         ).toSliver(),
         _SwitchSetting(
           title: 'Long press to zoom'.tl,
@@ -128,6 +209,8 @@ class _ReaderSettingsState extends State<ReaderSettings> {
             setState(() {});
             widget.onChanged?.call('enableLongPressToZoom');
           },
+          comicId: isEnabledSpecificSettings ? widget.comicId : null,
+          comicSource: isEnabledSpecificSettings ? widget.comicSource : null,
         ).toSliver(),
         SliverAnimatedVisibility(
           visible: appdata.settings['enableLongPressToZoom'] == true,
@@ -138,6 +221,8 @@ class _ReaderSettingsState extends State<ReaderSettings> {
               "press": "Press position".tl,
               "center": "Screen center".tl,
             },
+            comicId: isEnabledSpecificSettings ? widget.comicId : null,
+            comicSource: isEnabledSpecificSettings ? widget.comicSource : null,
           ),
         ),
         _SwitchSetting(
@@ -147,6 +232,8 @@ class _ReaderSettingsState extends State<ReaderSettings> {
           onChanged: () {
             widget.onChanged?.call('limitImageWidth');
           },
+          comicId: isEnabledSpecificSettings ? widget.comicId : null,
+          comicSource: isEnabledSpecificSettings ? widget.comicSource : null,
         ).toSliver(),
         if (App.isAndroid)
           _SwitchSetting(
@@ -155,6 +242,8 @@ class _ReaderSettingsState extends State<ReaderSettings> {
             onChanged: () {
               widget.onChanged?.call('enableTurnPageByVolumeKey');
             },
+            comicId: isEnabledSpecificSettings ? widget.comicId : null,
+            comicSource: isEnabledSpecificSettings ? widget.comicSource : null,
           ).toSliver(),
         _SwitchSetting(
           title: "Display time & battery info in reader".tl,
@@ -162,6 +251,8 @@ class _ReaderSettingsState extends State<ReaderSettings> {
           onChanged: () {
             widget.onChanged?.call("enableClockAndBatteryInfoInReader");
           },
+          comicId: isEnabledSpecificSettings ? widget.comicId : null,
+          comicSource: isEnabledSpecificSettings ? widget.comicSource : null,
         ).toSliver(),
         _SwitchSetting(
           title: "Show system status bar".tl,
@@ -169,6 +260,8 @@ class _ReaderSettingsState extends State<ReaderSettings> {
           onChanged: () {
             widget.onChanged?.call("showSystemStatusBar");
           },
+          comicId: isEnabledSpecificSettings ? widget.comicId : null,
+          comicSource: isEnabledSpecificSettings ? widget.comicSource : null,
         ).toSliver(),
         SelectSetting(
           title: "Quick collect image".tl,
@@ -184,6 +277,8 @@ class _ReaderSettingsState extends State<ReaderSettings> {
           help:
               "On the image browsing page, you can quickly collect images by sliding horizontally or vertically according to your reading mode"
                   .tl,
+          comicId: isEnabledSpecificSettings ? widget.comicId : null,
+          comicSource: isEnabledSpecificSettings ? widget.comicSource : null,
         ).toSliver(),
         _CallbackSetting(
           title: "Custom Image Processing".tl,
@@ -196,6 +291,8 @@ class _ReaderSettingsState extends State<ReaderSettings> {
           interval: 1,
           min: 1,
           max: 16,
+          comicId: isEnabledSpecificSettings ? widget.comicId : null,
+          comicSource: isEnabledSpecificSettings ? widget.comicSource : null,
         ).toSliver(),
         _SwitchSetting(
           title: "Show Page Number".tl,
@@ -203,6 +300,8 @@ class _ReaderSettingsState extends State<ReaderSettings> {
           onChanged: () {
             widget.onChanged?.call("showPageNumberInReader");
           },
+          comicId: isEnabledSpecificSettings ? widget.comicId : null,
+          comicSource: isEnabledSpecificSettings ? widget.comicSource : null,
         ).toSliver(),
       ],
     );
@@ -248,7 +347,7 @@ class __CustomImageProcessingState extends State<_CustomImageProcessing> {
               setState(() {});
             },
             child: Text("Reset".tl),
-          )
+          ),
         ],
       ),
       body: Column(
@@ -274,7 +373,7 @@ class __CustomImageProcessingState extends State<_CustomImageProcessing> {
                 ),
               ),
             ),
-          )
+          ),
         ],
       ),
     );
